project_7_web_app/requirements.txt
pandas
plotly_express
streamlit
sidetable 